requirements.txt
acres==0.4.1
aiofiles==24.1.0
altair==5.5.0
amadeus==12.0.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
Django==5.2.1
etelemetry==0.3.1
fastapi==0.115.12
filelock==3.18.0
frontend==0.0.3
gitdb==4.0.12
GitPython==3.1.44
groq==0.25.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
looseversion==1.3.0
lxml==5.4.0
MarkupSafe==3.0.2
narwhals==1.41.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.2.1
protobuf==6.31.1
prov==2.0.1
puremagic==1.29
pyarrow==20.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.0
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
scipy==1.15.3
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tools==1.0.2
tornado==6.5.1
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
watchdog==6.0.0